Paths and options moved to fixtures
