--- tests/fixtures/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "test-project",
-  "private": true,
-  "dependencies": {
-    "something-else": "latest"
-  },
-  "ember-addon": {
-    "paths": ["./lib/ember-super-button"]
-  },
-  "devDependencies": {
-    "ember-cli": "latest",
-    "ember-random-addon": "latest",
-    "ember-non-root-addon": "latest",
-    "ember-generated-with-export-addon": "latest",
-    "non-ember-thingy": "latest",
-    "ember-before-blueprint-addon": "latest",
-    "ember-after-blueprint-addon": "latest",
-    "ember-devDeps-addon": "latest",
-    "ember-addon-with-dependencies": "latest"
-  }
-}
