+++ tests/helpers/assert/file-exists.js
