+++ tests/helpers/assert/file-has-content.js
